apps refactoring wip
